Ulil fitur detail destinasi

+++ database/migrations/2026_04_24_155215_create_penilaian_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('wisata_id')->constrained('wisata')->cascadeOnDelete();
             $table->foreignId('kriteria_id')->constrained('kriteria')->cascadeOnDelete();
 
-            $table->float('nilai');
+            $table->decimal('nilai', 3, 2);
         });
     }
